+++ src/test/java/no/uib/inf112/desktop/CardTest.java
@@ -1,7 +1,7 @@
 package no.uib.inf112.desktop;
 
-import no.uib.inf112.player.Card;
-import no.uib.inf112.player.Movement;
+import no.uib.inf112.core.player.Card;
+import no.uib.inf112.core.player.Movement;
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
